+++ CHANGES.md
@@ -5,3 +5,4 @@
 * update for use with downloadable JSON files from GitHub Updater
 * add javascript to switch display from repos/sites
 * add update feedback messaging
+* create JSON storage directory in `wp-content/uploads/git-bulk-updater`